+++ js/modules/Zoom.js
@@ -8,9 +8,10 @@ class Zoom {
   // functions
   constructor(videoData) {
     this.scale = 1;
+    this.logScale = 0;
     this.scaleStep = 0.1;
-    this.minScale = 0.5;  // not accurate, actually slightly less
-    this.maxScale = 8;    // not accurate, actually slightly more
+    this.minScale = -1;  // 50% (log2(0.5) = -1)
+    this.maxScale = 3;   // 800% (log2(8) = 3)
     this.conf = videoData;
   }
 
@@ -18,41 +19,39 @@ class Zoom {
     this.scale = 1;
   }
 
-  zoomIn(){
-    this.scale += this.scaleStep;
-
-    if (this.scale >= this.maxScale) {
-      this.scale = this.maxScale;
-    }
-  }
-
-  zoomOut(){
-    this.scale -= this.scaleStep;
-
-    if (this.scale <= this.minScale) {
-      this.scale = this.minScale;
-    }
-    
-  }
-
   zoomStep(amount){
-    this.scale += amount;
+    this.logScale += amount;
 
-    if (this.scale <= this.minScale) {
-      this.scale = this.minScale;
+    if (this.logScale <= this.minScale) {
+      this.logScale = this.minScale;
     }
-    if (this.scale >= this.maxScale) {
-      this.scale = this.maxScale;
+    if (this.logScale >= this.maxScale) {
+      this.logScale = this.maxScale;
     }
+  
+    this.scale = Math.pow(2, this.logScale);
+
+    if (Debug.debug) {
+      console.log("[Zoom::zoomStep] changing zoom by", amount, ". New zoom level:", this.scale);
+    }
+
+    this.conf.restoreAr();
+    this.conf.announceZoom(this.scale);
   }
 
-  setZoom(scale){
-    if(scale < this.minScale) {
-      this.scale = this.minScale;
-    } else if (scale > this.maxScale) {
-      this.scale = this.maxScale;
-    } else {
-      this.scale = scale;
+  setZoom(scale, no_announce){
+    // NOTE: SCALE IS NOT LOGARITHMIC
+    if(scale < Math.pow(this.minScale)) {
+      scale = this.minScale;
+    } else if (scale > Math.pow(this.maxScale)) {
+      scale = this.maxScale;
+    }
+
+    this.scale = scale;
+
+    this.conf.restoreAr();
+    if (!no_announce) {
+      this.conf.announceZoom(this.scale);
     }
   }